Added 'main_template' to the catalogues.
Also added home button to cart.
