+++ public/css/catalogue.css
@@ -1,5 +1,5 @@
 *
-{
+/* {
     margin: 0;
     padding: 0;
 }
@@ -81,7 +81,7 @@ body
     margin-right: 10px;
     font-size: 17px;
     cursor: pointer;
-}
+} */
 .top-bar
 {
     width: 100%;
@@ -143,7 +143,7 @@ body
 .picture img:hover 
 {
     position:relative;
-    width:900px;
+    width:760px;
     height:auto;
     display:block;
 }
